--- src/main/java/com/zerobase/zbcharger/domain/charger/dto/admin/UpdateChargerRequest.java
@@ -1,10 +1,0 @@
-package com.zerobase.zbcharger.domain.charger.dto.admin;
-
-public record UpdateChargerRequest(
-    String chargerType,
-    String location,
-    String output,
-    String method
-) {
-
-}
