+++ src/main/java/com/zerobase/zbcharger/domain/charger/service/admin/ChargerService.java
@@ -7,11 +7,11 @@ import static com.zerobase.zbcharger.exception.constant.ErrorCode.STATION_NOT_FO
 
 import com.zerobase.zbcharger.domain.charger.dao.ChargerRepository;
 import com.zerobase.zbcharger.domain.charger.dao.StationRepository;
+import com.zerobase.zbcharger.domain.charger.dto.AddChargerRequest;
+import com.zerobase.zbcharger.domain.charger.dto.ChargerResponse;
 import com.zerobase.zbcharger.domain.charger.dto.ChargerSummary;
 import com.zerobase.zbcharger.domain.charger.dto.SearchChargerSummaryCondition;
-import com.zerobase.zbcharger.domain.charger.dto.admin.AddChargerRequest;
-import com.zerobase.zbcharger.domain.charger.dto.admin.ChargerResponse;
-import com.zerobase.zbcharger.domain.charger.dto.admin.UpdateChargerRequest;
+import com.zerobase.zbcharger.domain.charger.dto.UpdateChargerRequest;
 import com.zerobase.zbcharger.domain.charger.entity.Charger;
 import com.zerobase.zbcharger.exception.CustomException;
 import java.util.List;
@@ -40,7 +40,7 @@ public class ChargerService {
     }
 
     @Transactional(readOnly = true)
-    public Page<ChargerResponse> getChargerList(Pageable pageable) {
+    public Page<ChargerResponse> searchChargerList(Pageable pageable) {
         return chargerRepository.findAll(pageable).map(ChargerResponse::fromEntity);
     }
 
@@ -49,9 +49,8 @@ public class ChargerService {
         Charger charger = getChargerOrThrow(id);
 
         throwIfChargerDeleted(charger);
-
-        charger.update(request.chargerType(), request.location(), request.output(),
-            request.method());
+        
+        charger.update(request.chargerType(), request.method(), request.output());
     }
 
     @Transactional
